// src/app/omsorgspenger-utbetaling/types/EksisterendeOMPUTSoknaderActionKeys.ts
// eslint-disable-next-line import/prefer-default-export
export enum EksisterendeOMPUTSoknaderActionKeys {
    EKSISTERENDE_OMP_UT_SOKNADER_SET = 'PUNCH_EKSISTERENDE_OMP_UT_SOKNADER_SET',
    EKSISTERENDE_OMP_UT_SOKNADER_LOAD = 'PUNCH_EKSISTERENDE_OMP_UT_SOKNADER_LOAD',
    EKSISTERENDE_OMP_UT_SOKNADER_REQUEST_ERROR = 'PUNCH_EKSISTERENDE_OMP_UT_SOKNADER_REQUEST_ERROR',

    EKSISTERENDE_OMP_UT_SOKNAD_OPEN = 'PUNCH_EKSISTERENDE_OMP_UT_SOKNAD_OPEN',
    EKSISTERENDE_OMP_UT_SOKNAD_CLOSE = 'PUNCH_EKSISTERENDE_OMP_UT_SOKNAD_CLOSE',
    EKSISTERENDE_OMP_UT_SOKNAD_CHOOSE = 'PUNCH_EKSISTERENDE_OMP_UT_SOKNAD_CHOOSE',
    EKSISTERENDE_OMP_UT_SOKNAD_UNDO_CHOICE = 'PUNCH_EKSISTERENDE_OMP_UT_SOKNAD_UNDO_CHOICE',

    OMP_UT_SOKNAD_CREATE_REQUEST = 'PUNCH_EKSISTERENDE_OMP_UT_SOKNAD_CREATE_REQUEST',
    OMP_UT_SOKNAD_CREATE_ERROR = 'PUNCH_EKSISTERENDE_OMP_UT_SOKNAD_CREATE_ERROR',
    OMP_UT_SOKNAD_CREATE_SUCCESS = 'PUNCH_EKSISTERENDE_OMP_UT_SOKNAD_CREATE_SUCCESS',

    OMP_UT_SOKNADID_RESET = 'PUNCH_EKSTISTERENDE_OMP_UT_SOKNADID_RESET',
}